pyproject.toml
[project]
name = "terraform-google-tls-certificate"
version = "0.0.1"
description = "Test harness for Terraform/Tofu module"
readme = "README.md"
requires-python = ">=3.12"
# spell-checker: disable
dependencies = [
    "google-cloud-certificate-manager>=1.12.0",
    "google-cloud-compute>=1.43.0",
    "google-cloud-iam>=2.21.0",
    "google-cloud-resource-manager>=1.16.0",
    "google-cloud-secret-manager>=2.26.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
    "ty>=0.0.16",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
ignore = []
select = ["ALL"]
extend-select = ["I"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Allow assert and don't complain about implicit namespaces or too many arguements for tests.
"**/tests/**/*" = ["S", "INP", "PLR0913"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.coverage.paths]
source = ["tests"]

[tool.coverage.report]
show_missing = true

[tool.coverage.run]
branch = true

[tool.pytest]
addopts = [
    "--import-mode=importlib",
    "--cov=tests",
    "--cov-report=term-missing",
    "--numprocesses=auto",
    "--dist=loadfile",
]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s - %(levelname)s:%(name)s:%s(message)"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S%z"
testpaths = ["tests"]
